pyproject.toml
[project]
authors = [
    {name="Riccardo Cappuzzo", email ="riccardo.cappuzzo@inria.fr"}
]
readme = "README.md"
description = "A pipeline for benchmarking table augmentation methods for data lakes."
name = "retrieve-merge-predict"
version = "0.5"
requires-python = ">=3.10"
dependencies = [
    "pip",
    # "pandas==1.5.3",
    # "numpy==1.24.3",
    "woodwork==0.24.0",
    # "scikit-learn==1.2.2",
    "scipy==1.12.0",
    "matplotlib==3.7.1",
    "seaborn==0.12.2",
    "gitpython==3.1.31",
    "catboost==1.2",
    "tqdm==4.65.0",
    "jupyter",
    "featuretools==1.26.0",
    "pyarrow==12.0",
    "pre-commit==3.4.0",
    "toml",
    "datasketch==1.5.9",
    "chardet==3.0.4",
    "protobuf==3.20.3",
    "lazo-index-service>=0.7.0",
    "memory-profiler==0.61.0",
    "clevercsv==0.8.2",
    "polars-lts-cpu>=1.6.0",
    "skrub>=0.5.3",
    "pytabkit>=1.0.0",
]
license = { file = "LICENSE" }

[project.urls]
Homepage = "https://rcap107.github.io/retrieve-merge-predict/"
Repository = "https://github.com/rcap107/retrieve-merge-predict"

[tool.uv]
